+++ lib/jekyll/commands/page.rb
@@ -14,7 +14,7 @@ module Jekyll
 
       def self.options
         [
-          ['type', '-t TYPE', '--type TYPE', 'Specify the content type (file extension)'],
+          ['extension', '-x EXTENSION', '--extension EXTENSION', 'Specify the file extension'],
           ['layout', '-l LAYOUT', '--layout LAYOUT', "Specify the page layout"],
           ['force', '-f', '--force', 'Overwrite a page if it already exists']
         ]